+++ gomen/src/main/java/com/tuna/gomen/mapper/ApplicantMapper.java
@@ -10,14 +10,14 @@ import java.util.List;
 public interface ApplicantMapper {
 
     // 특정 팀 모집 게시글에 지원한 지원자 목록 조회
-    List<ApplicantDTO> findApplicantsByPostId(@Param("teamPostId") Long teamPostId);
+    List<ApplicantDTO> findApplicantsByPostId(@Param("teamPostId") Integer teamPostId);
 
     // 특정 유저가 지원한 모집 게시글 목록 조회
-    List<ApplicantDTO> findMyApplications(@Param("userId") Long userId);
+    List<ApplicantDTO> findMyApplications(@Param("userId") Integer userId);
 
 //    // 특정 유저가 지원한 모집 게시글에서 수락 여부 확인
 //    List<ApplicantDTO> findMyApplicationsWithStatus(@Param("userId") Long userId);
 
     // 특정 유저가 지원한 모집 게시글 상세 조회
-    List<ApplicantDTO> findMyApplicationsWithDetails(@Param("userId") Long userId);
+    List<ApplicantDTO> findMyApplicationsWithDetails(@Param("userId") Integer userId);
 }
